메인 로직들 객체로 리팩터링 - 맥뎁
