+++ src/main/kotlin/InputView.kt
@@ -1,0 +1,14 @@
+class InputView {
+
+    fun getInput(): GameString {
+        while(true) {
+            val inputWord = readLine()!!.trim()
+            try {
+                return GameString(inputWord)
+            } catch (exception: Exception) {
+                println(exception.message)
+                continue
+            }
+        }
+    }
+}
